made help better searching in cli and improvements to tui
