+++ .gitignore
@@ -3,6 +3,7 @@ test/
 summoner.json
 build/
 *.spec
+todo.md
 
 
 # Added by cargo
