+++ src/ui/mod.rs
@@ -1,2 +1,3 @@
 pub mod ui;
 pub mod app;
+pub mod keys;
